requirements.txt
Flask==3.1.0
Markdown==3.7
python-frontmatter==1.1.0
gunicorn==23.0.0
python-dateutil==2.9.0